Annotate some of the functions and classes.

+++ mcstatus/bedrock_status.py
@@ -13,10 +13,10 @@ class BedrockServerStatus:
         self.timeout = timeout
 
     @staticmethod
-    def parse_response(data: bytes, latency: int):
+    def parse_response(data: bytes, latency: float):
         data = data[1:]
         name_length = struct.unpack(">H", data[32:34])[0]
-        data = data[34 : 34 + name_length].decode().split(";")
+        data = data[34 : 34 + name_length].decode().split(";") # type: ignore
 
         try:
             map_ = data[7]
@@ -57,7 +57,7 @@ class BedrockServerStatus:
             data, _ = await stream.recv()
         finally:
             try:
-                stream.close()
+                stream.close() # type: ignore - stream may be unbound
             except BaseException:
                 pass
